Addressed a shortcoming where recursive calls would cause an IllegalStateException.
This was done by queuing inputs that come in while a machine is already evaluating an input. This should also allow concurrent calls (although that has not been considered as thoroughly.) This change required two main changes to the API. First, StateMachine#evaluateInput() no longer returns whether the input was ignored since the evaluation may be deferred. And second, it may throw an InterruptedException due to the use of a Semaphore to manage access to certain sections of code.